@@ -24,16 +24,16 @@ dependencies:
   flutter:
     sdk: flutter
 
-  google_fonts: ^2.1.0
-  responsive_builder: ^0.4.1
-  font_awesome_flutter: ^9.1.0 
+  google_fonts: ^6.2.1
+  responsive_builder: ^0.7.0
+  font_awesome_flutter: ^10.7.0 
   url_launcher: ^6.0.9
   flutter_feather_icons: ^2.0.0+1 
   animated_text_kit: ^4.2.1
-  visibility_detector: ^0.3.3
-  flutter_bloc: ^7.1.0
-  layout: ^0.5.1
-  freezed_annotation: ^0.14.0
+  visibility_detector: ^0.4.0+2
+  flutter_bloc: ^8.1.5
+  layout: ^1.0.5
+  freezed_annotation: ^2.4.1
   dartz: ^0.10.0
   get_it: ^7.2.0
   flutter_spinkit: ^5.0.0
@@ -47,7 +47,7 @@ dev_dependencies:
     sdk: flutter
 
   build_runner:
-  freezed: ^0.14.0
+  freezed: ^2.5.2
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
 
