@@ -3,7 +3,6 @@ import 'package:aerium/infrastructure/api/email_repository.dart';
 import 'package:bloc/bloc.dart';
 
 import 'package:freezed_annotation/freezed_annotation.dart';
-import 'package:meta/meta.dart';
 
 part 'email_event.dart';
 
@@ -24,24 +23,22 @@ class EmailBloc extends Bloc<EmailEvent, EmailState> {
   Stream<EmailState> mapEventToState(
     EmailEvent event,
   ) async* {
-    if (event is EmailEvent) {
-      yield EmailState.sendingEmail();
-      
-      final response = await _emailRepository.sendEmail(
-        name: event.name,
-        email: event.email,
-        subject: event.subject,
-        message: event.message,
-      );
+    yield EmailState.sendingEmail();
+    
+    final response = await _emailRepository.sendEmail(
+      name: event.name,
+      email: event.email,
+      subject: event.subject,
+      message: event.message,
+    );
 
-      yield* response.fold(
-        (failure) async* {
-          yield EmailState.failure();
-        },
-        (data) async* {
-          yield EmailState.emailSentSuccessFully();
-        },
-      );
+    yield* response.fold(
+      (failure) async* {
+        yield EmailState.failure();
+      },
+      (data) async* {
+        yield EmailState.emailSentSuccessFully();
+      },
+    );
     }
-  }
 }
